+++ Frontend/expensetracker/lib/src/auth/data/log_repo_impl.dart
@@ -1,65 +1,40 @@
-import 'package:expensetracker/src/auth/domain/models/login_response.dart';
 import 'package:expensetracker/src/auth/domain/models/sign_in.dart';
-import 'package:flutter/material.dart';
+import 'package:expensetracker/src/auth/domain/models/sign_up.dart';
+import 'package:expensetracker/src/auth/domain/repositories/log_repo.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:get/get.dart';
-import 'package:expensetracker/shared/cache/storage.dart';
 
-class LoginController extends GetxController with CacheManager {
-  TextEditingController nameController = TextEditingController();
-  TextEditingController passwordController = TextEditingController();
-
-  //safety method, getx should automatically dispose controllers when they are not used
+class LogRepoImplementation implements LogRepo {
   @override
-  void dispose() {
-    nameController.dispose();
-    passwordController.dispose();
-    super.dispose();
-  }
-
-  void clear() {
-    nameController.clear();
-    passwordController.clear();
-  }
-
-  Future<void> login() async {
-    final connect = GetConnect();
-    SignInInfo requestModel = SignInInfo(
-      name: nameController.text,
-      password: passwordController.text,
-    );
-
+  Future<Response> signIn(SignInInfo requestModel) async {
+    final GetConnect connect = GetConnect();
     try {
       final response = await connect.get(
         '${dotenv.env['API_URL']}/auth',
         headers: requestModel.toJson(),
       );
-      if (response.statusCode == 200) {
-        final responseModel = ResponseModel.fromJson(response.body);
-        saveToken(responseModel.message);
-        dispose();
-        Get.toNamed('/form');
-      } else {
-        throw Exception(response.body);
-      }
+      return response;
     } catch (error) {
-      showDialog(
-        context: Get.context!,
-        builder: (BuildContext context) {
-          return AlertDialog(
-            title: const Text("Erreur survenue"),
-            content: Text(error.toString()),
-            actions: [
-              TextButton(
-                onPressed: () {
-                  Get.back();
-                },
-                child: const Text("OK"),
-              ),
-            ],
-          );
-        },
-      );
+      throw Exception(error);
     }
+  }
+
+  @override
+  Future<Response> signUp(SignUpInfo info) async {
+    final GetConnect connect = GetConnect();
+    try {
+      final response = await connect.post(
+        '${dotenv.env['API_URL']}/user',
+        info.toJson(),
+      );
+      return response;
+    } catch (error) {
+      throw Exception(error);
+    }
+  }
+
+  @override
+  Future<SignInInfo> signInAuto(String token) async {
+    throw UnimplementedError();
   }
 }
