+++ Frontend/expensetracker/lib/src/auth/domain/models/sign_in.dart
@@ -4,7 +4,7 @@ import 'dart:convert';
 class SignInInfo {
   String name;
   String password;
-  
+
   SignInInfo({
     required this.name,
     required this.password,
